33Mail is not junk, it is a valid aliasing system
33mail is used to have multiple aliases, which redirect into the users personal inbox.

It's basically a privacy tool, and since it requires verification in the sign up process, it is rarely used by to create spam accounts. So I don't think it should be on this list. I know many (including myself) who have used this service for years to have different aliases for work and personal emails etc, while having everything land in the same inbox.

Happy to talk more about it,
Thanks,
Alicia 🚀